requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
jinja2
langchain
langchain-community
langchain-openai
langchain-pinecone
pypdf